chore: remove artifact and directory grouping
